+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 - (Bugfix) Check ArangoSync availability without checking healthiness
 - (Improvement) Add Anonymous Inspector mods
 - (Improvement) Do not check checksums for DeploymentReplicationStatus.IncomingSynchronization field values
+- (Improvement) Add ServerGroup details into ServerGroupSpec
 
 ## [1.2.20](https://github.com/arangodb/kube-arangodb/tree/1.2.20) (2022-10-25)
 - (Feature) Add action progress